Using the real Error(s)Body and GolemError in golem-client
