+++ golem-component-service/src/api/component.rs
@@ -14,7 +14,7 @@
 
 use crate::api::{ComponentError, Result};
 use futures_util::TryStreamExt;
-use golem_common::model::component::DefaultComponentOwner;
+use golem_common::model::component::{DefaultComponentOwner, VersionedComponentId};
 use golem_common::model::plugin::{
     DefaultPluginOwner, DefaultPluginScope, PluginInstallation, PluginInstallationCreation,
     PluginInstallationUpdate,
@@ -38,6 +38,7 @@ use poem_openapi::*;
 use std::collections::HashMap;
 use std::sync::Arc;
 use tracing::Instrument;
+use golem_common::model::error::{ErrorBody, ErrorsBody};
 
 pub struct ComponentApi {
     component_service: Arc<dyn ComponentService<DefaultComponentOwner>>,
