+++ golem-component-service/src/api/mod.rs
@@ -17,13 +17,13 @@ use golem_common::metrics::api::TraceErrorKind;
 use golem_common::SafeDisplay;
 use golem_component_service_base::service::component::ComponentError as ComponentServiceError;
 use golem_component_service_base::service::plugin::PluginError;
-use golem_service_base::model::{ErrorBody, ErrorsBody};
 use poem::endpoint::PrometheusExporter;
 use poem::error::ReadBodyError;
 use poem::Route;
 use poem_openapi::payload::Json;
 use poem_openapi::{ApiResponse, OpenApiService};
 use prometheus::Registry;
+use golem_common::model::error::{ErrorBody, ErrorsBody};
 
 pub mod component;
 pub mod healthcheck;
