+++ golem-shard-manager/src/error.rs
@@ -19,8 +19,8 @@ use std::fmt::Formatter;
 use golem_api_grpc::proto::golem;
 use golem_api_grpc::proto::golem::shardmanager::v1::shard_manager_error;
 use golem_common::metrics::api::TraceErrorKind;
+use golem_common::model::error::GolemError;
 use golem_common::retriable_error::IsRetriableError;
-use golem_service_base::model::GolemError;
 
 #[derive(thiserror::Error, Debug)]
 pub enum ShardManagerError {
