+++ golem-worker-service-base/src/api/error.rs
@@ -15,8 +15,8 @@
 use crate::service::component::ComponentServiceError;
 use crate::service::worker::WorkerServiceError;
 use golem_common::metrics::api::TraceErrorKind;
+use golem_common::model::error::{ErrorBody, ErrorsBody, GolemError, GolemErrorBody, GolemErrorUnknown};
 use golem_common::SafeDisplay;
-use golem_service_base::model::*;
 use poem_openapi::payload::Json;
 use poem_openapi::*;
 use tonic::Status;
