+++ golem-worker-service-base/src/gateway_binding/http_handler_binding.rs
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use golem_common::model::component::VersionedComponentId;
 use super::{IdempotencyKeyCompiled, WorkerNameCompiled};
-use golem_service_base::model::VersionedComponentId;
 use golem_wasm_ast::analysis::AnalysedExport;
 use rib::{Expr, RibError};
