+++ golem-worker-service-base/src/gateway_binding/worker_binding.rs
@@ -15,10 +15,10 @@
 use super::{IdempotencyKeyCompiled, InvocationContextCompiled, WorkerNameCompiled};
 use crate::gateway_rib_compiler::DefaultWorkerServiceRibCompiler;
 use crate::gateway_rib_compiler::WorkerServiceRibCompiler;
-use golem_service_base::model::VersionedComponentId;
 use golem_wasm_ast::analysis::AnalysedExport;
 use rib::{Expr, RibByteCode, RibError, RibInputTypeInfo, RibOutputTypeInfo, WorkerFunctionsInRib};
 use serde::{Deserialize, Serialize};
+use golem_common::model::component::VersionedComponentId;
 
 #[derive(Debug, Clone, PartialEq)]
 pub struct WorkerBinding {
