+++ golem-worker-service-base/src/service/worker/routing_logic.rs
@@ -29,10 +29,10 @@ use golem_api_grpc::proto::golem::workerexecutor::v1::worker_executor_client::Wo
 use golem_common::client::MultiTargetGrpcClient;
 use golem_common::model::RetryConfig;
 use golem_common::model::{Pod, ShardId, TargetWorkerId, WorkerId};
+use golem_common::model::error::{GolemError, GolemErrorInvalidShardId, GolemErrorUnknown};
 use golem_common::retriable_error::IsRetriableError;
 use golem_common::retries::get_delay;
 use golem_common::SafeDisplay;
-use golem_service_base::model::{GolemError, GolemErrorInvalidShardId, GolemErrorUnknown};
 use golem_service_base::service::routing_table::{HasRoutingTableService, RoutingTableError};
 
 use crate::service::worker::WorkerServiceError;
